update_status.py handler is singled out.
